+++ src/components/Story/Story.js
@@ -2,22 +2,10 @@ import React, { PropTypes } from 'react';
 import { FormattedRelative } from 'react-intl';
 import { Menu } from 'antd';
 import { Link } from 'react-router';
-import embedjs from 'embedjs';
-import PostFeedEmbed from './PostFeedEmbed';
+import StoryPreview from './StoryPreview';
 import StoryFooter from './StoryFooter';
-import BodyShort from './BodyShort';
-import { jsonParse } from '../../helpers/formatter';
-import { image } from '../../helpers/steemitLinks';
 import Avatar from '../Avatar';
 import Topic from '../Button/Topic';
-import {
-  getPositions,
-  isPostStartsWithAPicture,
-  isPostStartsWithAnEmbed,
-  isPostWithPictureBeforeFirstHalf,
-  isPostWithEmbedBeforeFirstHalf
-} from '../../post/Feed/PostFeedCardHelper';
-import { getHtml } from '../../post/Body';
 import './Story.less';
 
 const SubMenu = Menu.SubMenu;
@@ -72,51 +60,6 @@ class Story extends React.Component {
       onShareClick
     } = this.props;
 
-    const jsonMetadata = jsonParse(this.props.post.json_metadata);
-    let imagePath = ''; let bodyImg;
-    if (jsonMetadata.image && jsonMetadata.image[0]) {
-      imagePath = `https://steemitimages.com/600x800/${jsonMetadata.image[0]}`;
-    } else if ((bodyImg = post.body.match(image())) && bodyImg.length) { // eslint-disable-line no-cond-assign
-      imagePath = `https://steemitimages.com/600x800/${bodyImg[0]}`;
-    }
-    const embeds = embedjs.getAll(post.body);
-
-    const preview = {
-      text: () => (
-        <div key="text" className="Story__content__body">
-          <BodyShort body={post.body} />
-        </div>
-      ),
-
-      embed: () => (embeds && embeds[0]) && <PostFeedEmbed key="embed" embed={embeds[0]} />,
-
-      image: () => (
-        <Link key="image" to={post.url}>
-          <img alt="post" key={imagePath} src={imagePath} />
-        </Link>
-      ),
-    };
-
-    const htmlBody = getHtml(post.body, {}, 'text');
-    const tagPositions = getPositions(htmlBody);
-    const bodyData = [];
-
-    if (isPostStartsWithAPicture(tagPositions)) {
-      bodyData.push(preview.image());
-      bodyData.push(preview.text());
-    } else if (isPostStartsWithAnEmbed(tagPositions)) {
-      bodyData.push(preview.embed());
-      bodyData.push(preview.text());
-    } else if (isPostWithPictureBeforeFirstHalf(tagPositions)) {
-      bodyData.push(preview.text());
-      bodyData.push(preview.image());
-    } else if (isPostWithEmbedBeforeFirstHalf(tagPositions)) {
-      bodyData.push(preview.text());
-      bodyData.push(preview.embed());
-    } else {
-      bodyData.push(preview.text());
-    }
-
     return (
       <div className="Story">
         <Menu
@@ -151,7 +94,7 @@ class Story extends React.Component {
           <Link to={post.url}>
             <h2 className="Story__content__title">{post.title}</h2>
           </Link>
-          {bodyData}
+          <StoryPreview post={post} />
         </div>
         <div className="Story__footer">
           <StoryFooter
